Restructured the codebase
